@@ -1,53 +1,76 @@
 {
-  "name": "@2003scape/rsc-client",
-  "version": "1.1.0",
-  "description": "runescape classic web client",
-  "main": "./src/mudclient.js",
-  "scripts": {
-    "build-dev": "browserify -t bulkify -t brfs index.js > dist/index.bundle.js",
-    "build": "browserify -t bulkify -p tinyify index.js > dist/index.bundle.min.js",
-    "start": "st -p 1337 -nc -d dist"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/2003scape/rsc-client.git"
-  },
-  "keywords": [
-    "runescape",
-    "classic",
-    "rsc",
-    "client",
-    "mmorpg",
-    "game"
-  ],
-  "author": "2003Scape Team",
-  "license": "AGPL-3.0+",
-  "bugs": {
-    "url": "https://github.com/2003scape/rsc-client/issues"
-  },
-  "homepage": "https://github.com/2003scape/rsc-client#readme",
-  "devDependencies": {
-    "@ledgerhq/compressjs": "^1.3.2",
-    "alawmulaw": "^6.0.0",
-    "brfs": "^2.0.2",
-    "browserify": "^17.0.0",
-    "buffer": "^6.0.3",
-    "bulk-require": "^1.0.1",
-    "bulkify": "1.4.0",
-    "eslint": "^7.30.0",
-    "long": "^4.0.0",
-    "ndarray": "^1.0.19",
-    "pcm-player": "0.0.11",
-    "rand-token": "^1.0.1",
-    "sleep-promise": "^9.1.0",
-    "st": "^3.0.0",
-    "tga-js": "^1.1.1",
-    "tinyify": "^3.0.0"
-  },
-  "browserify": {
-    "transform": [
-      "bulkify",
-      "brfs"
-    ]
-  }
+	"name": "rsc-client",
+	"version": "1.1.0",
+	"description": "runescape classic web client",
+	"type": "module",
+	"scripts": {
+		"dev": "svelte-kit dev",
+		"build": "svelte-kit build",
+		"package": "svelte-kit package",
+		"preview": "svelte-kit preview",
+		"prepare": "svelte-kit sync",
+		"check": "svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
+		"lint": "prettier --check --plugin-search-dir=. . && eslint .",
+		"format": "prettier --write --plugin-search-dir=. .",
+		"legacy-build-dev": "browserify -t bulkify -t brfs index.js > dist/index.bundle.js",
+		"legacy-build": "browserify -t bulkify -p tinyify index.js > dist/index.bundle.min.js",
+		"legacy-start": "st -p 1337 -nc -d dist"
+	},
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/ty-ler/rsc-client.git"
+	},
+	"keywords": [
+		"runescape",
+		"classic",
+		"rsc",
+		"client",
+		"mmorpg",
+		"game"
+	],
+	"author": "2003Scape Team",
+	"license": "AGPL-3.0+",
+	"bugs": {
+		"url": "https://github.com/2003scape/rsc-client/issues"
+	},
+	"homepage": "https://github.com/2003scape/rsc-client#readme",
+	"devDependencies": {
+		"@ledgerhq/compressjs": "^1.3.2",
+		"@sveltejs/adapter-auto": "next",
+		"@sveltejs/kit": "next",
+		"@types/long": "^4.0.2",
+		"@typescript-eslint/eslint-plugin": "^5.27.0",
+		"@typescript-eslint/parser": "^5.27.0",
+		"alawmulaw": "^6.0.0",
+		"brfs": "^2.0.2",
+		"browserify": "^17.0.0",
+		"buffer": "^6.0.3",
+		"bulk-require": "^1.0.1",
+		"bulkify": "1.4.0",
+		"eslint": "^8.16.0",
+		"eslint-config-prettier": "^8.3.0",
+		"eslint-plugin-svelte3": "^4.0.0",
+		"long": "^4.0.0",
+		"ndarray": "^1.0.19",
+		"pcm-player": "0.0.11",
+		"prettier": "^2.6.2",
+		"prettier-plugin-svelte": "^2.7.0",
+		"rand-token": "^1.0.1",
+		"sleep-promise": "^9.1.0",
+		"st": "^3.0.0",
+		"svelte": "^3.44.0",
+		"svelte-check": "^2.7.1",
+		"svelte-preprocess": "^4.10.6",
+		"tga-js": "^1.1.1",
+		"tinyify": "^3.0.0",
+		"tslib": "^2.3.1",
+		"typescript": "^4.7.2"
+	},
+	"browserify": {
+		"transform": [
+			"bulkify",
+			"brfs"
+		]
+	}
 }
