@@ -1,7 +1,7 @@
-const serverOpcodes = require('../opcodes/server');
+import serverOpcodes from '../opcodes/server';
 
 module.exports = {
-    [serverOpcodes.APPEARANCE]: function () {
-        this.showAppearanceChange = true;
-    }
+	[serverOpcodes.APPEARANCE]: function () {
+		this.showAppearanceChange = true;
+	}
 };
