@@ -1,20 +1,20 @@
-const bulk = require('bulk-require');
+import bulk from 'bulk-require';
 
 function getPacketHandlers(mudclient) {
-    const handlers = bulk(__dirname, ['*.js']);
-    const packetMap = {};
+	const handlers = bulk(__dirname, ['*.js']);
+	const packetMap = {};
 
-    for (const [handlerName, handlerMap] of Object.entries(handlers)) {
-        if (/^_|index/.test(handlerName)) {
-            continue;
-        }
+	for (const [handlerName, handlerMap] of Object.entries(handlers)) {
+		if (/^_|index/.test(handlerName)) {
+			continue;
+		}
 
-        for (const [id, handler] of Object.entries(handlerMap)) {
-            packetMap[id] = handler.bind(mudclient);
-        }
-    }
+		for (const [id, handler] of Object.entries(handlerMap)) {
+			packetMap[id] = handler.bind(mudclient);
+		}
+	}
 
-    return packetMap;
+	return packetMap;
 }
 
 module.exports = getPacketHandlers;
