@@ -1,19 +1,19 @@
-const serverOpcodes = require('../opcodes/server');
+import serverOpcodes from '../opcodes/server';
 
 module.exports = {
-    [serverOpcodes.PRAYER_STATUS]: function (data, size) {
-        for (let i = 0; i < size - 1; i++) {
-            const on = data[i + 1] === 1;
+	[serverOpcodes.PRAYER_STATUS]: function (data, size) {
+		for (let i = 0; i < size - 1; i++) {
+			const on = data[i + 1] === 1;
 
-            if (!this.prayerOn[i] && on) {
-                this.playSoundFile('prayeron');
-            }
+			if (!this.prayerOn[i] && on) {
+				this.playSoundFile('prayeron');
+			}
 
-            if (this.prayerOn[i] && !on) {
-                this.playSoundFile('prayeroff');
-            }
+			if (this.prayerOn[i] && !on) {
+				this.playSoundFile('prayeroff');
+			}
 
-            this.prayerOn[i] = on;
-        }
-    }
+			this.prayerOn[i] = on;
+		}
+	}
 };
