@@ -1,26 +1,24 @@
-const PCMPlayer = require('pcm-player');
-const { mulaw } = require('alawmulaw');
+import PCMPlayer from 'pcm-player';
+import { mulaw } from 'alawmulaw';
 
 class StreamAudioPlayer {
-    constructor() {
-        this.player = new PCMPlayer({
-            encoding: '16bitInt',
-            channels: 1,
-            sampleRate: 8000
-        });
-    }
+	constructor() {
+		this.player = new PCMPlayer({
+			encoding: '16bitInt',
+			channels: 1,
+			sampleRate: 8000
+		});
+	}
 
-    stopPlayer() {
-        this.player.destroy();
-    }
+	stopPlayer() {
+		this.player.destroy();
+	}
 
-    writeStream(buffer, offset, length) {
-        const decoded = mulaw.decode(
-            new Uint8Array(buffer.slice(offset, offset + length))
-        );
+	writeStream(buffer, offset, length) {
+		const decoded = mulaw.decode(new Uint8Array(buffer.slice(offset, offset + length)));
 
-        this.player.feed(decoded);
-    }
+		this.player.feed(decoded);
+	}
 }
 
 module.exports = StreamAudioPlayer;
